skills/files/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyyaml==6.0.2
gitpython==3.1.43
python-dotenv==1.0.1
openpyxl==3.1.5
pandas==2.2.3
